refactor: Remove match_artists_on_wikidata
