+++ jobs/ml_jobs/artist_linkage/link_products_to_artists_from_scratch.py
@@ -13,7 +13,7 @@ from constants import (
     OFFER_CATEGORY_ID_KEY,
     PRODUCT_ID_KEY,
 )
-from match_artists_on_wikidata import load_wikidata
+from utils.loading import load_wikidata
 from utils.matching import create_artists_tables, match_artists_with_wikidata
 from utils.preprocessing_utils import filter_products, prepare_artist_names_for_matching
