+++ jobs/ml_jobs/artist_linkage/streamlits/st_investigate_syncho_backend.py
@@ -2,7 +2,6 @@ import pandas as pd
 import streamlit as st
 
 from constants import ARTIST_NAME_TO_FILTER
-from match_artists_on_wikidata import preprocess_artists
 from utils.preprocessing_utils import (
     clean_names,
     extract_first_artist,
@@ -12,6 +11,13 @@ from utils.preprocessing_utils import (
 DEFAULT_EXPANDER_STATE = False
 MERGE_COLUMNS = ["product_id", "artist_type"]
 st.set_page_config(layout="wide")
+
+
+def preprocess_artists(df: pd.DataFrame) -> pd.DataFrame:
+    return df.assign(
+        alias=lambda _df: _df.preprocessed_artist_name,
+        tmp_id=lambda df: df.index,
+    )
 
 
 @st.cache_data
